ARCHIVE/old_configs/compass-core/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
psycopg[binary]==3.2.9
PyYAML==6.0.1
python-dotenv==1.0.1
httpx==0.27.0
semver==3.0.2
